move influx metrics reporter to its own subproject
