+++ lib/dropwizard/metrics-influx/build.gradle
@@ -1,0 +1,5 @@
+apply from: rootProject.file('gradle/common_deps.gradle')
+dependencies {
+    compile project(':lib:dropwizard:base')
+    compile 'net.alchim31:metrics-influxdb:0.6.0'
+}
