+++ lib/dropwizard/metrics-influx/src/main/resources/META-INF/services/io.dropwizard.metrics.ReporterFactory
@@ -1,0 +1,1 @@
+com.wikia.dropwizard.metrics.influx.InfluxDbReporterFactory
