+++ lib/service-core/build.gradle
@@ -2,8 +2,8 @@ apply from: rootProject.file('gradle/pandora_dependencies.gradle')
 apply from: rootProject.file('gradle/publish_artifacts.gradle')
 
 dependencies {
-    compile 'net.alchim31:metrics-influxdb:0.6.0'
     compile project(':lib:dropwizard:base')
     compile project(':lib:dropwizard:hal')
     compile project(':lib:dropwizard:consul')
+    compile project(':lib:dropwizard:metrics-influx')
 }
