--- lib/service-core/src/main/resources/META-INF/services/io.dropwizard.metrics.ReporterFactory
@@ -1,1 +1,0 @@
-com.wikia.pandora.core.metrics.InfluxDbReporterFactory
